Refactor SpanKeyCreators to support non Stream based SpanKeyCreators
