+++ money-akka/src/main/scala/com/comcast/money/akka/stream/StreamSpanKeyCreators.scala
@@ -16,25 +16,11 @@
 
 package com.comcast.money.akka.stream
 
-import akka.stream.{ Attributes, FanOutShape, Inlet }
-import akka.stream.scaladsl.{ Flow, Source }
-import com.comcast.money.akka.SpanContextWithStack
+import akka.stream.scaladsl.{Flow, Source}
+import akka.stream.{Attributes, FanOutShape, Inlet}
+import com.comcast.money.akka.{SpanContextWithStack, SpanKeyCreator}
 
-import scala.reflect.{ ClassTag, classTag }
-
-/**
- * SpanKeyCreator and the traits that inherit from it use [[ClassTag]] and the Scala reflection
- * apis to get the names of the Inlet and/or Outlet types of a Stream Shape.
- * This is then used to create a key for the [[com.comcast.money.api.Span]] that is started in that [[akka.stream.Shape]]
- *
- * A Span key is used to determine the location in the code that a Span was made. It is important that it is unique
- * where possible. If your stream has many flows with the same outward type it may be best to use [[TracedFlow]]
- *
- * [[SpanKeyCreator]] is currently a work in progress
- */
-sealed trait SpanKeyCreator {
-  def nameOfType[T: ClassTag]: String = classTag[T].runtimeClass.getSimpleName
-}
+import scala.reflect.ClassTag
 
 /**
  * FanOutSpanKeyCreator provides users the capacity to name a [[FanOutShape]] based on the type of the [[FanOutShape]]
@@ -151,96 +137,4 @@ object InletSpanKeyCreator {
     new InletSpanKeyCreator[In] {
       override def inletToKey(inlet: Inlet[(In, SpanContextWithStack)])(implicit evIn: ClassTag[In]): String = toKey(inlet)
     }
-}
-
-/**
- * INTERNAL
- *
- * Default SpanKeyCreators to be used within the package as a way to allow for usage of the [[StreamTracingDSL]]
- * without having to define a implicit [[SpanKeyCreator]] for each [[akka.stream.Shape]] type used in the end users stream
- */
-
-private[akka] object DefaultSpanKeyCreators {
-
-  /**
-   * Returns a string to name the Span
-   *
-   * Example:
-   * Given In is type String
-   *
-   * returns "InletOfString"
-   */
-
-  object DefaultInletSpanKeyCreator {
-    def apply[In]: InletSpanKeyCreator[In] =
-      new InletSpanKeyCreator[In] {
-        override def inletToKey(inlet: Inlet[(In, SpanContextWithStack)])(implicit evIn: ClassTag[In]): String = s"InletOf${nameOfType[In]}"
-      }
-  }
-
-  /**
-   * Returns a string to name the Span
-   *
-   * Example:
-   * Given In is type String
-   *
-   * returns "FanInOfString"
-   */
-
-  object DefaultFanInSpanKeyCreator {
-    def apply[In]: FanInSpanKeyCreator[In] =
-      new FanInSpanKeyCreator[In] {
-        override def fanInInletToKey(inlet: Inlet[(In, SpanContextWithStack)])(implicit evIn: ClassTag[In]): String = s"FanInOf${nameOfType[In]}"
-      }
-  }
-
-  /**
-   * Returns a string to name the Span
-   *
-   * Example:
-   * Given In is type String
-   *
-   * returns "FanOutOfString"
-   */
-
-  object DefaultFanOutSpanKeyCreator {
-    def apply[In]: FanOutSpanKeyCreator[In] =
-      new FanOutSpanKeyCreator[In] {
-        override def fanOutToKey(fanOutShape: FanOutShape[(In, SpanContextWithStack)])(implicit evIn: ClassTag[In]): String = s"FanOutOf${nameOfType[In]}"
-      }
-  }
-
-  /**
-   * Returns a string to name the [[com.comcast.money.api.Span]] for the current [[Flow]]
-   *
-   * NB:
-   * Will default to the name provided in the [[Attributes]] of a [[Flow]]
-   * It is HIGHLY recommended to use [[Flow.named]] and to allow this to name the [[com.comcast.money.api.Span]] key
-   *
-   * Example:
-   * Given In is type String, Out is type String and there's no [[Attributes.Name]] defined for the [[Flow]]
-   *
-   * returns "StringToString"
-   */
-
-  object DefaultFlowSpanKeyCreator {
-    def apply[In]: FlowSpanKeyCreator[In] =
-      new FlowSpanKeyCreator[In] {
-        override def flowToKey[Out: ClassTag](flow: Flow[In, Out, _])(implicit evIn: ClassTag[In]): String =
-          Attributes.extractName(flow.traversalBuilder, s"${nameOfType[In]}To${nameOfType[Out]}")
-      }
-  }
-
-  /**
-   * Returns "Stream" a string to name the Span for the Stream
-   *
-   */
-
-  object DefaultSourceSpanKeyCreator {
-    def apply[In](): SourceSpanKeyCreator[In] =
-      new SourceSpanKeyCreator[In] {
-        override def sourceToKey(source: Source[In, _])(implicit evIn: ClassTag[In]): String = "Stream"
-      }
-  }
-
 }
